v3 compiles, but illegal mem access

+++ 4_1_parallel_search/v3/parallel_search.h
@@ -10,7 +10,7 @@
 
 #include <stdio.h>
 #include <inttypes.h>
-#include "../gpu_utils/util.h"
+#include "../../gpu_utils/util.h"
 
 int verbose = 0;
 
@@ -39,11 +39,11 @@ void _init(int argc, char ** argv);
 void _init_array(int with_file);
 
 /* ATOMIC FLAGS */
-// implementation here uses a byte for each flag : signaling completion for setting q and c
+// implementation here uses an int for each flag : signaling completion for setting q and c
 //			and an int for the number of iterations set by the thread setting r and l.
 __device__ int iter_flag;
-__device__ uint8_t * half_iter_signals;
-uint8_t * host_half_iter_signals_ptr;
+__device__ int * half_iter_signals;
+int * host_half_iter_signals_ptr;
 
 #include "cpu_search.h"
